feat: move types to their own file
